+++ voxygen/src/hud/buffs.rs
@@ -127,10 +127,11 @@ impl<'a> Widget for BuffsBar<'a> {
         .desc_text_color(TEXT_COLOR);
         if let BuffPosition::Bar = buff_position {
             let decayed_health = 1.0 - self.health.maximum() / self.health.base_max();
-            let show_health = (self.health.current() - self.health.maximum()).abs()
-                > Health::HEALTH_EPSILON
+            let show_health = self.global_state.settings.interface.always_show_bars
+                || (self.health.current() - self.health.maximum()).abs() > Health::HEALTH_EPSILON
                 || decayed_health > 0.0;
-            let show_energy = self.energy.current() != self.energy.maximum();
+            let show_energy = self.global_state.settings.interface.always_show_bars
+                || self.energy.current() != self.energy.maximum();
             let offset = if show_energy && show_health {
                 140.0
             } else if show_health || show_energy {