+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added a setting to always show health and energy bars
+
 ### Changed
 - Made dungeon tiers 3, 4, and 5 more common